pyproject.toml
[project]
name = "mfg-lifecycle-kb"
version = "0.1.0"
description = "Manufacturing Test Engineering Knowledge Base — NPI→MP lifecycle templates for NVIDIA GPU / DC-L6 / Automotive / Embedded products"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "jsonschema>=4.17.0",
    "python-frontmatter>=1.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[tool.uv]
package = false   # not a distributable package; uv manages the venv only

[tool.uv.sources]
# pin additional sources here if needed (private indexes, etc.)
